+++ lib/datahub.dart
@@ -9,6 +9,7 @@ export 'ioc.dart';
 export 'persistence.dart';
 export 'rest_client.dart';
 export 'services.dart';
+export 'storage.dart';
 export 'transfer_object.dart';
 export 'utils.dart';