+++ lib/src/broker/base_broker_client_service.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:datahub/datahub.dart';
 import 'package:dart_amqp/dart_amqp.dart';
 
-//TODO handle reconnects (channel based and connection based)
+@deprecated
 abstract class BaseBrokerClientService extends BaseService {
   final _log = resolve<LogService>();
   late final BrokerService _brokerService;