+++ lib/src/utils/encoding/transfer_codec.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:boost/boost.dart';
+import 'package:datahub/datahub.dart';
 
 import 'codec_exception.dart';
 
@@ -22,6 +23,38 @@ class TransferCodec<T> extends TypeCheck<T> {
     TransferCodec<DateTime>(_encodeDateTime, _decodeDateTime),
     TransferCodec<Duration>(_encodeDuration, _decodeDuration),
     TransferCodec<Uint8List>(_encodeUint8List, _decodeUint8List),
+    TransferCodec<List<String>>(encodeListTyped<List<String>, String>,
+        decodeListTyped<List<String>, String>),
+    TransferCodec<List<String?>>(encodeListTyped<List<String?>, String?>,
+        decodeListTyped<List<String?>, String?>),
+    TransferCodec<List<String>?>(encodeListTyped<List<String>?, String>,
+        decodeListTyped<List<String>?, String>),
+    TransferCodec<List<String?>?>(encodeListTyped<List<String?>?, String?>,
+        decodeListTyped<List<String?>?, String?>),
+    TransferCodec<List<int>>(
+        encodeListTyped<List<int>, int>, decodeListTyped<List<int>, int>),
+    TransferCodec<List<int?>>(
+        encodeListTyped<List<int?>, int?>, decodeListTyped<List<int?>, int?>),
+    TransferCodec<List<int>?>(
+        encodeListTyped<List<int>?, int>, decodeListTyped<List<int>?, int>),
+    TransferCodec<List<int?>?>(
+        encodeListTyped<List<int?>?, int?>, decodeListTyped<List<int?>?, int?>),
+    TransferCodec<List<double>>(encodeListTyped<List<double>, double>,
+        decodeListTyped<List<double>, double>),
+    TransferCodec<List<double?>>(encodeListTyped<List<double?>, double?>,
+        decodeListTyped<List<double?>, double?>),
+    TransferCodec<List<double>?>(encodeListTyped<List<double>?, double>,
+        decodeListTyped<List<double>?, double>),
+    TransferCodec<List<double?>?>(encodeListTyped<List<double?>?, double?>,
+        decodeListTyped<List<double?>?, double?>),
+    TransferCodec<List<bool>>(
+        encodeListTyped<List<bool>, bool>, decodeListTyped<List<bool>, bool>),
+    TransferCodec<List<bool?>>(encodeListTyped<List<bool?>, bool?>,
+        decodeListTyped<List<bool?>, bool?>),
+    TransferCodec<List<bool>?>(
+        encodeListTyped<List<bool>?, bool>, decodeListTyped<List<bool>?, bool>),
+    TransferCodec<List<bool?>?>(encodeListTyped<List<bool?>?, bool?>,
+        decodeListTyped<List<bool?>?, bool?>),
   ];
 
   static TransferCodec<T>? find<T>() =>